move count to decode module
