--- src/deli/cube/__init__.py
@@ -1,6 +1,0 @@
-"""init for deli.cube module"""
-
-from .count import corrected_count, dedup_count, raw_count
-
-
-__all__ = ["raw_count", "dedup_count", "corrected_count"]
