+++ src/deli/decode/count.py
